fix: move MobileTittle component
