+++ ui/applets/kit/src/components/index.ts
@@ -30,7 +30,6 @@ export { TextCopy } from "./TextCopy";
 export { Popover } from "./Popover";
 export { Pagination } from "./Pagination";
 export { JsonVisualizer } from "./JsonVisualizer";
-export { MobileTitle } from "./MobileTitle";
 
 /* ---------------------------------- Icons --------------------------------- */
