+++ ui/portal/web/src/pages/(app)/_app.account.$address.lazy.tsx
@@ -2,9 +2,9 @@ import { createLazyFileRoute } from "@tanstack/react-router";
 import { AccountExplorer } from "~/components/explorer/AccountExplorer";
 
 import type { Address } from "@left-curve/dango/types";
-import { MobileTitle } from "@left-curve/applets-kit";
 
 import { m } from "~/paraglide/messages";
+import { MobileTitle } from "~/components/foundation/MobileTitle";
 
 export const Route = createLazyFileRoute("/(app)/_app/account/$address")({
   component: AccountExplorerApplet,
@@ -15,11 +15,7 @@ function AccountExplorerApplet() {
 
   return (
     <div className="w-full flex flex-col">
-      <MobileTitle
-        action={() => history.go(-1)}
-        title={m["explorer.accounts.title"]()}
-        className="p-4 pb-0"
-      />
+      <MobileTitle title={m["explorer.accounts.title"]()} className="p-4 pb-0" />
       <AccountExplorer address={address as Address}>
         <AccountExplorer.NotFound />
         <AccountExplorer.Details />
