+++ ui/portal/web/src/pages/(app)/_app.block.$block.lazy.tsx
@@ -1,8 +1,8 @@
 import { m } from "~/paraglide/messages";
 import { createLazyFileRoute } from "@tanstack/react-router";
 
-import { MobileTitle } from "@left-curve/applets-kit";
 import { BlockExplorer } from "~/components/explorer/BlockExplorer";
+import { MobileTitle } from "~/components/foundation/MobileTitle";
 
 export const Route = createLazyFileRoute("/(app)/_app/block/$block")({
   component: BlockExplorerApplet,
@@ -13,11 +13,7 @@ function BlockExplorerApplet() {
 
   return (
     <div className="w-full flex flex-col">
-      <MobileTitle
-        action={() => history.go(-1)}
-        title={m["explorer.block.title"]()}
-        className="p-4 pb-0"
-      />
+      <MobileTitle title={m["explorer.block.title"]()} className="p-4 pb-0" />
       <BlockExplorer height={block}>
         <BlockExplorer.Skeleton />
         <BlockExplorer.NotFound />
