+++ ui/portal/web/src/pages/(app)/_app.notifications.lazy.tsx
@@ -1,8 +1,9 @@
-import { createLazyFileRoute } from "@tanstack/react-router";
+import { useMediaQuery } from "@left-curve/applets-kit";
 
-import { MobileTitle, useMediaQuery } from "@left-curve/applets-kit";
 import { Notifications } from "~/components/notifications/Notifications";
+import { MobileTitle } from "~/components/foundation/MobileTitle";
 
+import { createLazyFileRoute } from "@tanstack/react-router";
 import { m } from "~/paraglide/messages";
 
 export const Route = createLazyFileRoute("/(app)/_app/notifications")({
@@ -14,7 +15,7 @@ function NotificationApplet() {
   return (
     <div className="w-full md:max-w-[50rem] mx-auto flex flex-col gap-4 p-4 pt-6 mb-16">
       <div className="flex items-center justify-between">
-        <MobileTitle action={() => history.go(-1)} title={m["notifications.title"]()} />
+        <MobileTitle title={m["notifications.title"]()} />
       </div>
       <Notifications
         maxNotifications={isMd ? 10 : 5}
