+++ ui/portal/web/src/pages/(app)/_app.settings.lazy.tsx
@@ -1,25 +1,23 @@
+import { useAccount, useSessionKey } from "@left-curve/store";
 import { createLazyFileRoute, useRouter } from "@tanstack/react-router";
+import { useApp } from "~/hooks/useApp";
 
 import {
-  IconButton,
-  IconChevronDown,
   IconFormatNumber,
   IconInfo,
   IconLanguage,
   IconMobile,
   IconUser,
-  MobileTitle,
   Select,
   useMediaQuery,
 } from "@left-curve/applets-kit";
-import { useAccount, useSessionKey } from "@left-curve/store";
 import { Modals } from "~/components/modals/RootModal";
+import { MobileTitle } from "~/components/foundation/MobileTitle";
 import { KeyManagement } from "~/components/settings/KeyManagement";
-import { useApp } from "~/hooks/useApp";
+import { SessionCountdown } from "~/components/settings/SessionCountdown";
+
 import { m } from "~/paraglide/messages";
 import { getLocale, locales, setLocale } from "~/paraglide/runtime";
-
-import { SessionCountdown } from "~/components/settings/SessionCountdown";
 
 export const Route = createLazyFileRoute("/(app)/_app/settings")({
   component: SettingsComponent,
@@ -27,7 +25,6 @@ export const Route = createLazyFileRoute("/(app)/_app/settings")({
 
 function SettingsComponent() {
   const { isLg } = useMediaQuery();
-  const { history } = useRouter();
   const { isConnected, username } = useAccount();
   const { showModal, changeSettings, settings } = useApp();
   const { formatNumberOptions } = settings;
@@ -35,7 +32,7 @@ function SettingsComponent() {
 
   return (
     <div className="w-full md:max-w-[50rem] mx-auto flex flex-col gap-5 p-4 pt-6 mb-16">
-      <MobileTitle action={() => history.go(-1)} title={m["settings.title"]()} />
+      <MobileTitle title={m["settings.title"]()} />
       {session ? (
         <div className="rounded-xl bg-rice-25 shadow-card-shadow flex flex-col w-full px-2 py-4">
           <h3 className="h4-bold text-gray-900 px-2 pb-4">{m["settings.session.title"]()}</h3>
