+++ ui/portal/web/src/pages/(app)/_app.swap.lazy.tsx
@@ -1,9 +1,9 @@
-import { IconButton, IconChevronDown, MobileTitle } from "@left-curve/applets-kit";
 import { createLazyFileRoute, useNavigate } from "@tanstack/react-router";
 
 import { m } from "~/paraglide/messages";
 
 import { SimpleSwap } from "~/components/dex/SimpleSwap";
+import { MobileTitle } from "~/components/foundation/MobileTitle";
 
 export const Route = createLazyFileRoute("/(app)/_app/swap")({
   component: SwapApplet,
@@ -19,7 +19,7 @@ function SwapApplet() {
 
   return (
     <div className="w-full md:max-w-[25rem] mx-auto flex flex-col p-4 pt-6 gap-4 min-h-[100svh] md:min-h-fit">
-      <MobileTitle action={() => navigate({ to: "/" })} title={m["dex.convert.title"]()} />
+      <MobileTitle title={m["dex.convert.title"]()} />
       <SimpleSwap pair={{ from, to }} onChangePair={onChangePair}>
         <SimpleSwap.Header />
         <SimpleSwap.Form />
