+++ ui/portal/web/src/pages/(app)/_app.tx.$txHash.lazy.tsx
@@ -1,7 +1,8 @@
 import { m } from "~/paraglide/messages";
-import { MobileTitle } from "@left-curve/applets-kit";
+
 import { createLazyFileRoute } from "@tanstack/react-router";
 import { TransactionExplorer } from "~/components/explorer/TransactionExplorer";
+import { MobileTitle } from "~/components/foundation/MobileTitle";
 
 export const Route = createLazyFileRoute("/(app)/_app/tx/$txHash")({
   component: TransactionExplorerApplet,
@@ -12,11 +13,7 @@ function TransactionExplorerApplet() {
 
   return (
     <div className="w-full flex flex-col">
-      <MobileTitle
-        action={() => history.go(-1)}
-        title={m["explorer.txs.title"]()}
-        className="p-4 pb-0"
-      />
+      <MobileTitle title={m["explorer.txs.title"]()} className="p-4 pb-0" />
       <TransactionExplorer txHash={txHash}>
         <TransactionExplorer.NotFound />
         <TransactionExplorer.Details />
